Polished Menus/Panels and added level selection panel
This updated polished some menus, like adding a singleplayer clarification to the tutorial selection game mode, as well as reformatting the end game screen for tutorial mode to incloude an option that takes the user directly to the controls panel. I also added a level selection panel after selecting 2P mode so that users may choose one of three levels to play each other one. This feature still needs images for the levels, as well as levels 2 and 3 to be made.