Fixed Merge conflicts, fixed bullets no appearing, fixed game text, fixed audio playback upon death
This is a BIG commit, I had to resolve the merge conflicts in the unity scenes, which resulted in having to delete the generative level feature. I was able to fix the bullet feature and the game text and the audio playback upon death, but was unable to resolve the generative feature. this took a lot of work and a lot of time, sorry its all in one commit but my goal was to get everything in working order for submission tonight.